Detect decimal/hex confusion
Add a test for hex digits that are actually decimal (0-9 only).

The chances that a hex nibble is 0-9 10/16, so the math for
number of hex digits that must all be 0-9 to hit the false
positive rate is:

(10/16)^n = 1/2^60

n = log(1/2^60)/log(10/16) = 88.5
... or 89 hex digits which is 45 bytes.

Thanks to Emin Gün Sirer for the suggestion.